[FIX] Adapted Queries to match DB Schema. [ADD] .gitignore to ignore file containing DB information. [ADD] helper.php to contain functions that are used frequently.
